# doc-expert/src/api/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"servers":[{"url":"http://127.0.0.1:8000","description":"Local dev server"}],"paths":{"/chat/":{"post":{"tags":["chat"],"summary":"Send Message","operationId":"chat-send_message","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatInSchema"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat-stream/":{"post":{"tags":["chat"],"summary":"Send Message Stream","operationId":"chat-send_message_stream","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatInSchema"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/source/":{"post":{"tags":["source"],"summary":"Get Source","operationId":"source-get_source","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetSourceRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetSourceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["root"],"summary":"Main","operationId":"root-main","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ChatInSchema":{"title":"ChatInSchema","required":["query","current_url","history"],"type":"object","properties":{"query":{"title":"Query","type":"string"},"current_url":{"title":"Current Url","type":"string"},"history":{"title":"History","type":"array","items":{"$ref":"#/components/schemas/ChatMessageSchema"}}}},"ChatMessageSchema":{"title":"ChatMessageSchema","required":["message","sender"],"type":"object","properties":{"message":{"title":"Message","type":"string"},"sender":{"title":"Sender","type":"string"}}},"ChatResponse":{"title":"ChatResponse","required":["answer","source_content","source_metadata"],"type":"object","properties":{"answer":{"title":"Answer","type":"string"},"source_content":{"title":"Source Content","type":"string"},"source_metadata":{"title":"Source Metadata","type":"object"}}},"GetSourceRequest":{"title":"GetSourceRequest","required":["url"],"type":"object","properties":{"url":{"title":"Url","type":"string"}}},"GetSourceResponse":{"title":"GetSourceResponse","type":"object","properties":{"source_id":{"title":"Source Id","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}